Added ft_strcmp to libft, checker implementation complete, added make rules to check memory leaks of the checker, standard case, and error case tests.

+++ src/lib/checker/src/lib/libft/src/ft_string/ft_strncmp.c
@@ -11,7 +11,7 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-size_t	ft_strncmp(const char *ptr_find, const char *ptr_str, size_t length)
+int	ft_strncmp(const char *ptr_find, const char *ptr_str, size_t length)
 {
 	size_t			index;
 	unsigned char	*aux1;
